requirements.txt
beautifulsoup4
deep_translator
deepface
Flask
google-genai
opencv-python
playsound
psutil
pyautogui
pypdf2
python-dotenv
pyttsx3
pywhatkit
screen_brightness_control
selenium
SpeechRecognition
webdriver-manager
youtube_transcript_api
tf-keras
mem0ai
google-generativeai
pgvector
psycopg2
flask-sqlalchemy
flask-login
werkzeug
Flask-Login
db-sqlite3